# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "MediaCategoryCreateRequest": {
        "properties": {
          "description": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          }
        },
        "type": "object"
      },
      "MediaCategoryDto": {
        "properties": {
          "description": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "MediaCategoryUpdateRequest": {
        "properties": {
          "description": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "MediaDto": {
        "properties": {
          "averageRating": {
            "format": "double",
            "type": "number"
          },
          "bucketName": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/MediaCategoryDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "durationSeconds": {
            "format": "int32",
            "type": "integer"
          },
          "favouriteId": {
            "format": "int64",
            "type": "integer"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "isFavourite": {
            "type": "boolean"
          },
          "isPremium": {
            "type": "boolean"
          },
          "mediaObject": {
            "$ref": "#/components/schemas/S3ObjectDto"
          },
          "name": {
            "type": "string"
          },
          "pictureObject": {
            "$ref": "#/components/schemas/S3ObjectDto"
          },
          "requiresPremiumSubscription": {
            "type": "boolean"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "ERROR"
            ],
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "MediaFileUpdateRequest": {
        "properties": {
          "bucketName": {
            "minLength": 1,
            "type": "string"
          },
          "categoryId": {
            "format": "int64",
            "type": "integer"
          },
          "description": {
            "type": "string"
          },
          "durationSeconds": {
            "format": "int32",
            "type": "integer"
          },
          "file": {
            "format": "binary",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "picture": {
            "format": "binary",
            "type": "string"
          },
          "requiresPremiumSubscription": {
            "type": "boolean"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "ERROR"
            ],
            "type": "string"
          },
          "tagIds": {
            "items": {
              "format": "int64",
              "type": "integer"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "S3ObjectDto": {
        "properties": {
          "baseUrl": {
            "type": "string"
          },
          "bucketName": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "objectUri": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Keycloak-issued access token. Two realms are accepted, the mobile realm and the admin realm; the token's issuer selects which. Endpoints under /admin additionally require the ADMIN role.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Yogida Team",
      "url": "https://github.com/yogida"
    },
    "description": "REST API for the Yogida meditation platform. \nProvides presigned Cloudflare R2 streaming URLs and a catalog of available meditation media grouped by bucket.",
    "title": "Meditation Service API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/media/{id}": {
      "delete": {
        "description": "Removes the DB record and deletes the associated S3 object. S3 deletion failure is logged but does not affect the response.",
        "operationId": "deleteMedia",
        "parameters": [
          {
            "description": "Media ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Media deleted"
          },
          "404": {
            "description": "Media not found"
          }
        },
        "summary": "Delete media record and its S3 object",
        "tags": [
          "Admin — Media"
        ]
      },
      "get": {
        "operationId": "getAdminMediaById",
        "parameters": [
          {
            "description": "Media ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaDto"
                }
              }
            },
            "description": "Media found"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaDto"
                }
              }
            },
            "description": "Media not found"
          }
        },
        "summary": "Get media by ID",
        "tags": [
          "Admin — Media"
        ]
      },
      "put": {
        "description": "Updates metadata. If a new file is provided and the object key differs, uploads it to S3 and removes the old object.",
        "operationId": "updateMedia",
        "parameters": [
          {
            "description": "Media ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/MediaFileUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaDto"
                }
              }
            },
            "description": "Media updated"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaDto"
                }
              }
            },
            "description": "Media not found"
          }
        },
        "summary": "Update media record",
        "tags": [
          "Admin — Media"
        ]
      }
    },
...
    "/media-categories": {
      "get": {
        "operationId": "listMediaCategories",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MediaCategoryDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Media categories retrieved successfully"
          }
        },
        "summary": "Get all media categories",
        "tags": [
          "Media Categories"
        ]
      },
      "post": {
        "operationId": "createMediaCategory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaCategoryCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaCategoryDto"
                }
              }
            },
            "description": "Media category created successfully"
          }
        },
        "summary": "Create a new media category",
        "tags": [
          "Media Categories"
        ]
      }
    },
    "/media-categories/{id}": {
      "delete": {
        "operationId": "deleteMediaCategory",
        "parameters": [
          {
            "description": "Media category ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Media category deleted successfully"
          },
          "404": {
            "description": "Media category not found"
          }
        },
        "summary": "Delete a media category",
        "tags": [
          "Media Categories"
        ]
      },
      "get": {
        "operationId": "getMediaCategoryById",
        "parameters": [
          {
            "description": "Media category ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaCategoryDto"
                }
              }
            },
            "description": "Media category found"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaCategoryDto"
                }
              }
            },
            "description": "Media category not found"
          }
        },
        "summary": "Get media category by ID",
        "tags": [
          "Media Categories"
        ]
      },
      "put": {
        "operationId": "updateMediaCategory",
        "parameters": [
          {
            "description": "Media category ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaCategoryUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaCategoryDto"
                }
              }
            },
            "description": "Media category updated successfully"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaCategoryDto"
                }
              }
            },
            "description": "Media category not found"
          }
        },
        "summary": "Update an existing media category",
        "tags": [
          "Media Categories"
        ]
      }
    },
    "/media/stream": {
      "post": {
        "description": "Returns a time-limited presigned URL for streaming the media object.",
        "operationId": "getMediaStreamUrl",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/S3ObjectDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "url": "https://r2.example.com/bucket/media.mp3?X-Amz-Signature=..."
                  },
                  "type": "object"
                }
              }
            },
            "description": "Presigned URL generated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "error": "Forbidden",
                    "message": "Not entitled to access this media",
                    "path": "/api/media/stream",
                    "status": 403,
                    "timestamp": "2026-07-24T10:00:00"
                  },
                  "type": "object"
                }
              }
            },
            "description": "User is not entitled to access this media"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Media not found"
          }
        },
        "summary": "Get presigned streaming URL",
        "tags": [
          "Media"
        ]
      }
    },
    "/media/{id}": {
      "get": {
        "description": "Returns a single media item by its database ID.",
        "operationId": "getMediaById",
        "parameters": [
          {
            "description": "Media ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaDto"
                }
              }
            },
            "description": "Media found"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaDto"
                }
              }
            },
            "description": "Media not found"
          }
        },
        "summary": "Get media by ID",
        "tags": [
          "Media"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Current environment",
      "url": "/api"
    }
  ],
...
}